refactor: move gitDiff files functionality out of zipUtil
