+++ packages/core/src/codewhisperer/models/constants.ts
@@ -926,3 +926,7 @@ export const testGenExcludePatterns = [
     '**/*.deb',
     '**/*.model',
 ]
+
+export const isFileAnalysisScope = (scope: CodeAnalysisScope) => {
+    return scope === CodeAnalysisScope.FILE_AUTO || scope === CodeAnalysisScope.FILE_ON_DEMAND
+}
