+++ packages/core/src/codewhisperer/util/zipUtil.ts
@@ -25,12 +25,11 @@ import {
     ProjectSizeExceededError,
 } from '../models/errors'
 import { FeatureUseCase } from '../models/constants'
-import { ChildProcess, ChildProcessOptions } from '../../shared/utilities/processUtils'
 import { ProjectZipError } from '../../amazonqTest/error'
-import { removeAnsi } from '../../shared/utilities/textUtilities'
 import { normalize } from '../../shared/utilities/pathUtils'
 import { ZipStream } from '../../shared/utilities/zipStream'
 import { getTextContent } from '../../shared/utilities/vsCodeUtils'
+import { getGitDiffContent } from '../../amazonq/util/gitDiffUtils'
 
 export interface ZipMetadata {
     rootDir: string
@@ -48,13 +47,6 @@ export const ZipConstants = {
     gitignoreFilename: '.gitignore',
     knownBinaryFileExts: ['.class'],
     codeDiffFilePath: 'codeDiff/code.diff',
-}
-
-interface GitDiffOptions {
-    projectPath: string
-    projectName: string
-    filePath?: string
-    scope?: CodeWhispererConstants.CodeAnalysisScope
 }
 
 export class ZipUtil {
@@ -78,15 +70,10 @@ export class ZipUtil {
         return CodeWhispererConstants.projectScanPayloadSizeLimitBytes
     }
 
-    public reachSizeLimit(size: number, scope: CodeWhispererConstants.CodeAnalysisScope): boolean {
-        if (
-            scope === CodeWhispererConstants.CodeAnalysisScope.FILE_AUTO ||
-            scope === CodeWhispererConstants.CodeAnalysisScope.FILE_ON_DEMAND
-        ) {
-            return size > this.getFileScanPayloadSizeLimitInBytes()
-        } else {
-            return size > this.getProjectScanPayloadSizeLimitInBytes()
-        }
+    protected reachSizeLimit(size: number, scope: CodeWhispererConstants.CodeAnalysisScope): boolean {
+        return CodeWhispererConstants.isFileAnalysisScope(scope)
+            ? size > this.getFileScanPayloadSizeLimitInBytes()
+            : size > this.getProjectScanPayloadSizeLimitInBytes()
     }
 
     public willReachSizeLimit(current: number, adding: number): boolean {
@@ -230,164 +217,10 @@ export class ZipUtil {
         filePath?: string,
         scope?: CodeWhispererConstants.CodeAnalysisScope
     ) {
-        let gitDiffContent = ''
-        for (const projectPath of projectPaths) {
-            const projectName = path.basename(projectPath)
-            // Get diff content
-            gitDiffContent += await this.executeGitDiff({
-                projectPath,
-                projectName,
-                filePath,
-                scope,
-            })
-        }
+        const gitDiffContent = await getGitDiffContent(projectPaths, filePath)
         if (gitDiffContent) {
             zip.writeString(gitDiffContent, ZipConstants.codeDiffFilePath)
         }
-    }
-
-    private async getGitUntrackedFiles(projectPath: string): Promise<string | undefined> {
-        const checkNewFileArgs = ['ls-files', '--others', '--exclude-standard']
-        const checkProcess = new ChildProcess('git', checkNewFileArgs)
-
-        try {
-            let output = ''
-            await checkProcess.run({
-                rejectOnError: true,
-                rejectOnErrorCode: true,
-                onStdout: (text) => {
-                    output += text
-                },
-                spawnOptions: {
-                    cwd: projectPath,
-                },
-            })
-            return output
-        } catch (err) {
-            getLogger().warn(`Failed to check if file is new: ${err}`)
-            return undefined
-        }
-    }
-
-    private async generateNewFileDiff(projectPath: string, projectName: string, relativePath: string): Promise<string> {
-        let diffContent = ''
-
-        const gitArgs = [
-            'diff',
-            '--no-index',
-            `--src-prefix=a/${projectName}/`,
-            `--dst-prefix=b/${projectName}/`,
-            '/dev/null', // Use /dev/null as the old file
-            relativePath,
-        ]
-
-        const childProcess = new ChildProcess('git', gitArgs)
-        const runOptions: ChildProcessOptions = {
-            rejectOnError: false,
-            rejectOnErrorCode: false,
-            onStdout: (text) => {
-                diffContent += text
-                getLogger().verbose(removeAnsi(text))
-            },
-            onStderr: (text) => {
-                getLogger().error(removeAnsi(text))
-            },
-            spawnOptions: {
-                cwd: projectPath,
-            },
-        }
-
-        try {
-            await childProcess.run(runOptions)
-            return diffContent
-        } catch (err) {
-            getLogger().warn(`Failed to run diff command: ${err}`)
-            return ''
-        }
-    }
-
-    private async generateHeadDiff(projectPath: string, projectName: string, relativePath?: string): Promise<string> {
-        let diffContent = ''
-
-        const gitArgs = [
-            'diff',
-            'HEAD',
-            `--src-prefix=a/${projectName}/`,
-            `--dst-prefix=b/${projectName}/`,
-            ...(relativePath ? [relativePath] : []),
-        ]
-
-        const childProcess = new ChildProcess('git', gitArgs)
-
-        const runOptions: ChildProcessOptions = {
-            rejectOnError: true,
-            rejectOnErrorCode: true,
-            onStdout: (text) => {
-                diffContent += text
-                getLogger().verbose(removeAnsi(text))
-            },
-            onStderr: (text) => {
-                getLogger().error(removeAnsi(text))
-            },
-            spawnOptions: {
-                cwd: projectPath,
-            },
-        }
-
-        try {
-            await childProcess.run(runOptions)
-            return diffContent
-        } catch (err) {
-            getLogger().warn(`Failed to run command \`${childProcess.toString()}\`: ${err}`)
-            return ''
-        }
-    }
-
-    private async executeGitDiff(options: GitDiffOptions): Promise<string> {
-        const { projectPath, projectName, filePath, scope } = options
-        const isProjectScope = scope === CodeWhispererConstants.CodeAnalysisScope.PROJECT
-
-        const untrackedFilesString = await this.getGitUntrackedFiles(projectPath)
-        const untrackedFilesArray = untrackedFilesString?.trim()?.split('\n')?.filter(Boolean)
-
-        if (isProjectScope && untrackedFilesArray && !untrackedFilesArray.length) {
-            return await this.generateHeadDiff(projectPath, projectName)
-        }
-
-        let diffContent = ''
-
-        if (isProjectScope) {
-            diffContent = await this.generateHeadDiff(projectPath, projectName)
-
-            if (untrackedFilesArray) {
-                const untrackedDiffs = await Promise.all(
-                    untrackedFilesArray.map((file) => this.generateNewFileDiff(projectPath, projectName, file))
-                )
-                diffContent += untrackedDiffs.join('')
-            }
-        } else if (!isProjectScope && filePath) {
-            const relativeFilePath = path.relative(projectPath, filePath)
-
-            const newFileDiff = await this.generateNewFileDiff(projectPath, projectName, relativeFilePath)
-            diffContent = this.rewriteDiff(newFileDiff)
-        }
-        return diffContent
-    }
-
-    private rewriteDiff(inputStr: string): string {
-        const lines = inputStr.split('\n')
-        const rewrittenLines = lines.slice(0, 5).map((line) => {
-            line = line.replace(/\\\\/g, '/')
-            line = line.replace(/("a\/[^"]*)/g, (match, p1) => p1)
-            line = line.replace(/("b\/[^"]*)/g, (match, p1) => p1)
-            line = line.replace(/"/g, '')
-
-            return line
-        })
-        const outputLines = [...rewrittenLines, ...lines.slice(5)]
-        const outputStr = outputLines.join('\n')
-
-        return outputStr
     }
 
     protected async processSourceFiles(
